rearrange componets to their own structure so cds_usergen.pages is the the default view
