+++ packages/cds-usergen/cds_usergen/components/__init__.py
@@ -1,1 +1,3 @@
 from .validated_inputs import ValidatedTextInput, ValidatedInputInt
+from .components import *
+from .class_components import *
